Fix trielog shipping issue (#68)
This fix enables the recreation of the account in a new leaf index in the event of a self-destruct and recreate within the same block. Prior to this fix, Shomei did not perform this operation, which could result in pointing to the old version (leaf index) of an account after the recreate. This led to an issue where we did not have an empty storage and ended up retrieving the storage from before the self-destruct, which is not the expected behavior. Now, with this fix, we ensure that a new leaf index is created for the recreated account, providing an accurate and empty storage state.